+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.0
+
+**Breaking**
+* Replaced blockedTimeframes from TimeframeSelector with overlayDatas
+
+**Changes**
+* Added OverlayData to allow customisation of overlays
+
+
 ## 0.0.1
 
 * First release of timeframe_selector